+++ commons/src/main/AndroidManifest.xml
@@ -50,7 +50,12 @@
             android:exported="false"
             android:label="@string/third_party_licences"
             android:parentActivityName="com.simplemobiletools.commons.activities.AboutActivity" />
-        <activity android:name=".activities.TestManageBlockedNumbers" />
+
+        <activity
+            android:name=".activities.ManageBlockedNumbersActivity"
+            android:configChanges="orientation"
+            android:exported="false"
+            android:label="@string/manage_blocked_numbers" />
 
         <receiver
             android:name=".receivers.SharedThemeReceiver"
