+++ commons/src/main/kotlin/com/simplemobiletools/commons/activities/FAQActivity.kt
@@ -3,13 +3,12 @@ package com.simplemobiletools.commons.activities
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.core.view.WindowCompat
 import com.simplemobiletools.commons.compose.extensions.AdjustNavigationBarColors
-import com.simplemobiletools.commons.compose.extensions.TransparentSystemBars
 import com.simplemobiletools.commons.compose.screens.FAQScreen
 import com.simplemobiletools.commons.compose.theme.AppThemeSurface
 import com.simplemobiletools.commons.helpers.APP_FAQ
@@ -19,9 +18,8 @@ import kotlinx.collections.immutable.toImmutableList
 class FAQActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        WindowCompat.setDecorFitsSystemWindows(window, false)
+        enableEdgeToEdge()
         setContent {
-            TransparentSystemBars()
             AppThemeSurface {
                 var canScroll by remember { mutableStateOf<Boolean?>(null) }
                 AdjustNavigationBarColors(canScroll)
