+++ commons/src/main/kotlin/com/simplemobiletools/commons/compose/settings/scaffold/SettingsScaffoldExtensions.kt
@@ -5,13 +5,16 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.TopAppBarScrollBehavior
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.lerp
 import androidx.compose.ui.platform.LocalLayoutDirection
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.simplemobiletools.commons.compose.extensions.onEventValue
+import com.simplemobiletools.commons.compose.extensions.rememberWindowInsetsController
 import com.simplemobiletools.commons.compose.theme.isNotLitWell
 import com.simplemobiletools.commons.compose.theme.isSurfaceLitWell
 import com.simplemobiletools.commons.extensions.getColoredMaterialStatusBarColor
@@ -19,10 +22,8 @@ import com.simplemobiletools.commons.extensions.getContrastColor
 
 @Composable
 internal fun SystemUISettingsScaffoldStatusBarColor(scrolledColor: Color) {
-    val systemUiController = rememberSystemUiController()
-    SideEffect {
-        systemUiController.setStatusBarColor(Color.Transparent, darkIcons = scrolledColor.isNotLitWell())
-    }
+    val insetController = rememberWindowInsetsController()
+    insetController.isAppearanceLightStatusBars = scrolledColor.isNotLitWell()
 }
 
 @Composable
