+++ gradle/libs.versions.toml
@@ -27,15 +27,15 @@ recyclerviewFastscroller = "5a95285b1f"
 #rtlViewpager
 rtlViewpager = "940f12724f"
 #Compose
-composeActivity = "1.8.0-alpha06"
-compose = "1.6.0-alpha03"
-composeCompiler = "1.5.1"
-composeMaterial3 = "1.2.0-alpha05"
-accompanist = "0.33.0-alpha"
+composeActivity = "1.8.0-alpha07"
+compose = "1.6.0-alpha04"
+composeCompiler = "1.5.2"
+composeMaterial3 = "1.2.0-alpha06"
+accompanist = "0.33.1-alpha"
 #Glide
 glide = "4.16.0"
 #Room
-room = "2.6.0-alpha03"
+room = "2.6.0-beta01"
 #ezVcard
 ezVcard = "0.11.3"
 #gson
@@ -87,8 +87,6 @@ compose-ui = { module = "androidx.compose.ui:ui", version.ref = "compose" }
 compose-runtime = { module = "androidx.compose.runtime:runtime", version.ref = "compose" }
 compose-uiTooling-debug = { module = "androidx.compose.ui:ui-tooling", version.ref = "compose" }
 compose-uiTooling-preview = { module = "androidx.compose.ui:ui-tooling-preview", version.ref = "compose" }
-#Accompanist
-accompanist-systemuicontroller = { module = "com.google.accompanist:accompanist-systemuicontroller", version.ref = "accompanist" }
 #Glide
 glide = { module = "com.github.bumptech.glide:glide", version.ref = "glide" }
 glide-compiler = { module = "com.github.bumptech.glide:compiler", version.ref = "glide" }
@@ -125,9 +123,6 @@ compose-preview = [
 room = [
     "androidx-room-ktx",
     "androidx-room-runtime",
-]
-accompanist = [
-    "accompanist-systemuicontroller",
 ]
 lifecycle = [
     "androidx-lifecycle-compose",
