+++ samples/src/main/kotlin/com/simplemobiletools/commons/samples/activities/MainActivity.kt
@@ -3,7 +3,7 @@ package com.simplemobiletools.commons.samples.activities
 import android.content.Intent
 import android.os.Bundle
 import com.simplemobiletools.commons.activities.BaseSimpleActivity
-import com.simplemobiletools.commons.activities.TestManageBlockedNumbers
+import com.simplemobiletools.commons.activities.ManageBlockedNumbersActivity
 import com.simplemobiletools.commons.dialogs.BottomSheetChooserDialog
 import com.simplemobiletools.commons.extensions.appLaunched
 import com.simplemobiletools.commons.extensions.toast
@@ -41,7 +41,7 @@ class MainActivity : BaseSimpleActivity() {
             launchAbout()
         }
         binding.manageBlockedNumbers.setOnClickListener {
-            startActivity(Intent(this, TestManageBlockedNumbers::class.java))
+            startActivity(Intent(this, ManageBlockedNumbersActivity::class.java))
         }
     }
